+++ src/main/java/net/etylop/immersivefarming/event/IFEvents.java
@@ -4,23 +4,27 @@ import net.etylop.immersivefarming.ImmersiveFarming;
 import net.etylop.immersivefarming.block.IFBlocks;
 import net.etylop.immersivefarming.block.custom.Soil;
 import net.etylop.immersivefarming.utils.CropSavedData;
+import net.etylop.immersivefarming.utils.ModTags;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Registry;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.item.HoeItem;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.level.block.CropBlock;
-import net.minecraft.world.level.block.GrassBlock;
+import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.event.entity.player.BonemealEvent;
-import net.minecraftforge.event.entity.player.UseHoeEvent;
+import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.Event;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import static net.etylop.immersivefarming.utils.IFFunctions.isBlockTillable;
+import java.util.Collections;
+import java.util.Optional;
 
 
 public class IFEvents {
@@ -28,15 +32,14 @@ public class IFEvents {
     public static class ForgeEvents {
 
         @SubscribeEvent
-        public static void onHoeUse(UseHoeEvent event) {
-            UseOnContext context = event.getContext();
-            Level level = context.getLevel();
-            if (isBlockTillable(level, context.getClickedPos())) {
-                level.setBlock(context.getClickedPos(), IFBlocks.SOIL.get().defaultBlockState(), 3);
-                event.getPlayer().getMainHandItem().hurtAndBreak(1, event.getPlayer(), (val) -> {
-                    val.broadcastBreakEvent(event.getPlayer().getUsedItemHand());
-                });
-            }
+        public static void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
+            Block target = event.getWorld().getBlockState(event.getPos()).getBlock();
+            if (Registry.BLOCK.getHolderOrThrow(Registry.BLOCK.getResourceKey(target).get()).is(ModTags.Blocks.TILLABLE_BLOCK) &&
+                !event.getWorld().isClientSide() &&
+                event.getPlayer().getMainHandItem().getItem() instanceof HoeItem) {
+
+                event.getWorld().setBlock(event.getPos(), IFBlocks.SOIL.get().defaultBlockState(), 3);
+                event.getPlayer().getMainHandItem().hurtAndBreak(1, event.getPlayer(), (val) -> val.broadcastBreakEvent(event.getPlayer().getUsedItemHand()));            }
         }
 
         @SubscribeEvent
@@ -52,8 +55,10 @@ public class IFEvents {
             event.setCanceled(true);
         }
 
-        @SubscribeEvent
+        @SubscribeEvent(priority = EventPriority.LOW)
         public static void onCropGrowth(BlockEvent.CropGrowEvent.Pre event) {
+            if (event.getResult() == Event.Result.DENY) return;
+
             LevelAccessor level = event.getWorld();
             BlockState cropBlock =  level.getBlockState(event.getPos());
 
@@ -65,8 +70,12 @@ public class IFEvents {
                 return;
             }
 
+            if (!(cropBlock.getBlock() instanceof CropBlock || cropBlock.getBlock() instanceof StemBlock)) {
+                return;
+            }
+
             if (canCropGrow((Level) event.getWorld(), event.getPos())) {
-                BlockState soilBlock =  event.getWorld().getBlockState(event.getPos().below());
+                BlockState soilBlock = event.getWorld().getBlockState(event.getPos().below());
                 int fertilization = soilBlock.getValue(Soil.FERTILITY);
 
                 if (fertilization>=1) {
@@ -84,18 +93,49 @@ public class IFEvents {
                         event.setResult(Event.Result.DENY);
                     }
                 }
-                if (cropBlock.getValue(CropBlock.AGE)==CropBlock.MAX_AGE-1 && !(event.getResult()==Event.Result.DENY)) {
-                    level.setBlock(event.getPos().below(), soilBlock.setValue(Soil.FERTILITY, 0), 2);
+            }
+        }
+
+        @SubscribeEvent
+        public static void onCropGrowthPost(BlockEvent.CropGrowEvent.Post event) {
+            BlockState cropBlock = event.getState();
+            LevelAccessor level = event.getWorld();
+            BlockState soilBlock = event.getWorld().getBlockState(event.getPos().below());
+
+            if (!(soilBlock.getBlock() instanceof Soil)) {
+                return;
+            }
+
+            if (cropBlock.getBlock() instanceof CropBlock || cropBlock.getBlock() instanceof StemBlock) {
+                Optional<Integer> ageOptional = cropBlock.getValues().entrySet().stream()
+                        .filter(entry -> "age".equals(entry.getKey().getName()) && entry.getValue() instanceof Integer)
+                        .map(entry -> (Integer) entry.getValue())
+                        .findFirst();
+
+                Optional<Integer> maxAgeOptional = cropBlock.getValues().entrySet().stream()
+                        .filter(entry -> "age".equals(entry.getKey().getName()) && entry.getKey() instanceof IntegerProperty)
+                        .map(entry -> Collections.max(((IntegerProperty) entry.getKey()).getPossibleValues()))
+                        .findFirst();
+
+                int age = ageOptional.orElse(0);
+                int maxAge = maxAgeOptional.orElse(10);
+
+                if (age == maxAge) {
                     CropSavedData cropData = getCropSavedData((Level) event.getWorld());
-                    if (cropData!=null) {
+                    if (cropData != null) {
                         cropData.insertCrop((Level) level, event.getPos());
                         cropData.setDirty();
                     }
+                    level.setBlock(event.getPos().below(), soilBlock.setValue(Soil.FERTILITY, 0), 2);
                     event.setResult(Event.Result.ALLOW);
                 }
-                return;
             }
-            event.setResult(Event.Result.DENY);
+
+            if (cropBlock.getBlock() instanceof AttachedStemBlock) {
+                if (Math.random() < 0) { // TODO add config
+                    level.setBlock(event.getPos(), Blocks.AIR.defaultBlockState(), 2);
+                }
+            }
         }
 
         private static boolean canCropGrow(Level level, BlockPos pos) {