+++ changelog.md
@@ -10,6 +10,16 @@
   * Change probability of disease contamination
 
 
+##### Version 1.18.2-0.2.3
+
+* Fixed default growth for non crops/stems
+
+
+##### Version 1.18.2-0.2.2
+
+* Fixed crash when crops have a different maxAge
+
+
 ##### Version 1.18.2-0.2.1
 
 * Growth speed fix
@@ -29,7 +39,6 @@
 * Changed the sprinkler models
 * Engineer's manual integration
 * Reduced crop disease chance
-
 
 ##### Version 1.18.2-0.1